# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nelle Agent API",
    "version": "1.0.0",
    "description": "Local-first AI agent server. Generated from zod contracts."
  },
  "components": {
    "schemas": {
      "NelleError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1
          },
          "message": {
            "type": "string",
            "minLength": 1
          },
          "detail": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "logRef": {
            "type": "string"
          }
        },
        "required": ["code", "message"],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/runtime": {
      "get": {
        "summary": "GET /api/runtime",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/NelleError"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/install/stream": {
      "post": {
        "summary": "POST /api/runtime/install/stream",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/NelleError"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/update/stream": {
      "post": {
        "summary": "POST /api/runtime/update/stream",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/NelleError"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/start": {
      "post": {
        "summary": "POST /api/runtime/start",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/NelleError"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/stop": {
      "post": {
        "summary": "POST /api/runtime/stop",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/NelleError"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/uninstall": {
      "post": {
        "summary": "POST /api/runtime/uninstall",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/NelleError"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/runtime/logs": {
      "get": {
        "summary": "GET /api/runtime/logs",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "$ref": "#/components/schemas/NelleError"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}